Update README.md
Updated links and descriptions where applicable for the following:
[CellularAutomaton] - redirected the link to the new area with .cs file
ColorHSVDrawer - redirected the link to the new area with .cs file
MeshFilterExtension - redirected the link to the new area with .cs file
ProceduralToolkitMenu - redirected the link to the new area with .cs file
RulesetDrawer - redirected the link to the new area with .cs file

Tom Kelly (2014) Unwritten procedural modeling with the straight skeleton. Link - replaced the 404 Twak link with University of Glasgow Link taken from the Twak Source site

The Clipper library.  Link - replaced the 404 angusj link with Clipper 2 link